requirements.txt
streamlit
streamlit-authenticator
google-generativeai
SQLAlchemy
supabase
praw
openai==0.28